pyproject.toml
[project]
name = "langchain-job-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "deepagents>=0.1.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "langchain-openai>=0.1.0",
    "playwright>=1.42.0",
    "langchain-community>=0.0.10",
    "curl-cffi>=0.6.0",
    "seleniumbase>=4.24.0",
]
